% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElbows.R
\name{getElbows}
\alias{getElbows}
\title{Given a decreasingly sorted vector, return the given number of elbows}
\usage{
getElbows(dat, n = 3, threshold = FALSE, plot = TRUE, main = "")
}
\arguments{
\item{dat}{a input vector (e.g. a vector of standard deviations), or a input feature matrix}

\item{n}{the number of returned elbows}

\item{threshold}{either FALSE or a number. If threshold is a number, then all the elements in d that are not larger than the threshold will be ignored.}

\item{plot}{logical. When T, it depicts a scree plot with highlighted elbows}

\item{main}{a string of the plot title}
}
\value{
a vector of length \eqn{n}
}
\description{
Given a decreasingly sorted vector, return the given number of elbows
}
\references{
Zhu, Mu and Ghodsi, Ali (2006), Automatic dimensionality selection from
the scree plot via the use of profile likelihood, Computational
Statistics & Data Analysis, Volume 51 Issue 2, pp 918-930, November, 2006.
}
\author{
Youngser Park <youngser@jhu.edu>
}
