% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmase.R
\name{gmmase}
\alias{gmmase}
\title{Run spectral clustering on a (possibly directed) (possibly weighted) graph.}
\usage{
gmmase(g, dmax = 20, elb = 1, embed = "ASE", clustering = "GMM",
  use.ptr = TRUE, verbose = TRUE)
}
\arguments{
\item{g}{a graph in \code{igraph} format}

\item{dmax}{maximum dimension for embedding}

\item{elb}{an index for elbow}

\item{embed}{either \code{ASE} or \code{LSE}, spectral embedding method}

\item{clustering}{either \code{GMM} or \code{Kmeans}, clustering method}

\item{use.ptr}{boolean to determine whether to perform pass-to-rank or not, default is \code{TRUE}}

\item{verbose}{boolean to determine whether to display an intermediate fitting progress status of \code{mclust} or not, default is \code{TRUE}}
}
\value{
\code{g} the largest connected component of the input graph

\code{mc} clustering output object

\code{Y} labels for the clustering
}
\description{
It does
1. do a pass-to-rank for a weighted graph (\eqn{PTR}, no-op for an unweighted graph),
2. do a graph spectral embedding (\code{ASE} or \code{LSE}) with a _diagonal augmentation_,
3. do a _dimension reduction_ (\eqn{ZG}) and merge left and right vectors (no-op for an undirected graph),
4. cluster vertices (\eqn{GMM} or \eqn{Kmeans}).
}
\examples{
library(igraph)
data(g)
E(g)$weight <- runif(ecount(g), 1, 5) # add random edge weights
Y <- gmmase(g, dmax=20, use.ptr=TRUE, embed="ASE", clustering="Kmeans")

}
\references{
D.L. Sussman, M. Tang, D.E. Fishkind, and C.E. Priebe,
A consistent adjacency spectral embedding for stochastic blockmodel graphs,
Journal of the American Statistical Association, Vol. 107, No. 499, pp. 1119-1128, 2012.
}
\author{
Youngser Park <youngser@jhu.edu>
}
