% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar.R
\name{nonpar}
\alias{nonpar}
\title{Nonparametric two-sample testing using kernel-based test statistic}
\usage{
nonpar(Xhat1, Xhat2, sigma = 0.5)
}
\arguments{
\item{Xhat1}{a \eqn{n} x \eqn{d} matrix}

\item{Xhat2}{a \eqn{n} x \eqn{d} matrix}

\item{sigma}{a bandwidth for the Gaussian kernel}
}
\value{
\code{T} A scalar value \eqn{T} such that \eqn{T} is near 0 if the rows of
\eqn{X} and \eqn{Y} are from the same distribution and \eqn{T} far from 0 if the rows of
\eqn{X} and \eqn{Y} are from different distribution.
}
\description{
This is a simple implementation of the kernel-based test statistic for the nonparametric
two-sample testing problem of given \eqn{X_1, X_2, \dots, X_n} i.i.d. \eqn{F} and
\eqn{Y_1, Y_2, \dots, Y_m} i.i.d. \eqn{G}, test the null hypothesis of \eqn{F = G} against
the alternative hypothesis of \eqn{F \not = G}. The test statistic is based on embedding
\eqn{F} and \eqn{G} into a reproducing kernel Hilbert space and then compute a distance between
the resulting embeddings. For this primitive, the Hilbert space is associated with the
Gaussian kernel.
}
\author{
Youngser Park <youngser@jhu.edu>
}
