% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptr.R
\name{ptr}
\alias{ptr}
\title{Run pass-to-rank on a weighted graph.}
\usage{
ptr(g)
}
\arguments{
\item{g}{a graph in \code{igraph} format}
}
\description{
It extract edge weight \eqn{W} and divide \eqn{2*R / (|E|+1)} where \eqn{R} is the rank of \eqn{W} and \eqn{|E|} is the number of edges.
}
\author{
Youngser Park <youngser@jhu.edu>
}
